+++ CHANGELOG.md
@@ -9,6 +9,10 @@ All notable changes to this project will be documented in this file. The format 
 - Showing loading spinner when fetching KPI/KeyResult progress
 - Moved widgets around for a better UX experience for mobile users
 
+### Fixed
+
+- functions: config to active slackbot or not actually works
+
 ## [2.5.2] 2021-10-19
 
 ### Fixed